renamed user folder to userpage for consistency
